rename: 상품 상세 _store 폴더 이름 변경 (#59)
